+++ ddp-workspace/projects/ddp-dsm-ui/src/app/FON/fon.component.ts
@@ -1,5 +1,11 @@
-import {Component} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {Title} from '@angular/platform-browser';
+import {SessionService} from "../services/session.service";
+import {SessionMementoService} from "ddp-sdk";
+import {Auth} from "../services/auth.service";
+import {takeUntil} from "rxjs/operators";
+import {Subject} from "rxjs";
+import {TranslateService} from "@ngx-translate/core";
 
 @Component({
   selector: 'app-fon',
@@ -39,8 +45,36 @@ import {Title} from '@angular/platform-browser';
   `]
 })
 
-export class FonComponent {
-  constructor(private title: Title) {
+export class FonComponent implements OnInit, OnDestroy {
+  unsubscribe = new Subject();
+  private readonly SESSION_KEY: string = 'session_key';
+
+  constructor(private title: Title,
+              private sessionService: SessionService,
+              private dssSessionService: SessionMementoService,
+              private authService: Auth,
+              private translateService: TranslateService) {
     title.setTitle('Fon');
+  }
+
+  ngOnInit() {
+    this.authService.dsmToken.pipe(takeUntil(this.unsubscribe))
+      .subscribe(token => token && this.setDssSession(token))
+
+    this.translateService.setDefaultLang('en');
+    this.translateService.use('en')
+  }
+
+  ngOnDestroy() {
+    this.unsubscribe.next(null);
+  }
+
+  private setDssSession(dsmToken: string): void {
+    const accessToken = null;
+    const userGuid = null;
+    const locale = 'en';
+    const expiresAtInSeconds: number = +this.sessionService.getTokenExpiration();
+    // set DSS Session partially
+    this.dssSessionService.setSession(accessToken, dsmToken, userGuid, locale, expiresAtInSeconds);
   }
 }
