+++ ddp-workspace/projects/ddp-dsm-ui/src/app/FON/pages/activities/activity/activity.component.ts
@@ -1,7 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Params} from '@angular/router';
 import {SessionMementoService} from 'ddp-sdk';
-import {ComponentService} from '../../../../services/component.service';
 
 @Component({
   selector: 'app-activity',
@@ -17,7 +16,6 @@ import {ComponentService} from '../../../../services/component.service';
 
 export class ActivityComponent implements OnInit {
   activityGuid: string;
-  selectedRealm: string = localStorage.getItem(ComponentService.MENU_SELECTED_REALM);
 
   constructor(private router: ActivatedRoute, private sessionService: SessionMementoService) {
   }
