--- ddp-workspace/projects/ddp-dsm-ui/src/app/services/logging.service.spec.ts
@@ -1,15 +1,0 @@
-// import { TestBed, inject } from '@angular/core/testing';
-
-// import { LoggingDsmService } from './logging.service';
-
-describe('LoggingDsmService', () => {
-  // beforeEach(() => {
-  //   TestBed.configureTestingModule({
-  //     providers: [LoggingDsmService]
-  //   });
-  // });
-  //
-  // it('should be created', inject([LoggingDsmService], (service: LoggingDsmService) => {
-  //   expect(service).toBeTruthy();
-  // }));
-});
