+++ ddp-workspace/projects/ddp-sdk/src/lib/ddp.module.ts
@@ -1,4 +1,4 @@
-import { ErrorHandler, NgModule } from '@angular/core';
+import {ErrorHandler, ModuleWithProviders, NgModule} from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
@@ -7,7 +7,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { Title } from '@angular/platform-browser';
 import { A11yModule } from '@angular/cdk/a11y';
 // ngx-translate
-import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import {TranslateLoader, TranslateModule, TranslateService} from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { NGXTranslateService } from './services/internationalization/ngxTranslate.service';
 // CookieService
@@ -523,4 +523,14 @@ export function createTranslateLoader(http: HttpClient): TranslateHttpLoader {
     ]
 })
 export class DdpModule {
+    static forDSM(): ModuleWithProviders<DdpModule> {
+        return {
+            ngModule: DdpModule,
+            providers: [
+                SessionMementoService,
+                StackdriverErrorReporterService,
+                TranslateService
+            ]
+        }
+    }
 }
